pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tf-trend"
version = "0.9.0"
description = "Systematic trend-following research and backtesting framework (starter repo)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Engineer Investor (@egr_investor)"}]
dependencies = [
    "numpy==1.26.4",
    "pandas>=2.2.2",
    "matplotlib==3.8.2",
    "pyyaml==6.0.1",
    "scipy==1.11.4",
    "tabulate==0.9.0",
    "rich==13.7.0",
    "yfinance==0.2.36",
    "joblib==1.3.2",
]

[project.urls]
Homepage = "https://github.com/engineerinvestor/systematic-trend-following-with-managed-futures"
Repository = "https://github.com/engineerinvestor/systematic-trend-following-with-managed-futures"
"Issue Tracker" = "https://github.com/engineerinvestor/systematic-trend-following-with-managed-futures/issues"
Twitter = "https://x.com/egr_investor"

[project.scripts]
tf = "tf.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
